feat(types): create a custom AnomalyDetector type to enable the auto-complete
Sometimes, having 1 single class as a type, mostly if it's an interface class, will disable the auto-complete and disturb the linters because the child classes will not be visible in the background. In Python 3.12, we have the flexibility to create a simple custom types that lets us still have the accurate type for an object while being able to enjoy the auto-complete and not-so-crazy liter e.g. MyPy.