feat(tests): create DataGenerator to generate simple testable data for our detectors
DataGenerator class provides the functionalities to generate data with random values from 0 to 1_000 or 10_000 or -1_000 or 10_000 to 1_000 or 10_000 enhanced with customable amount of anomaly e.g. currently I used 0.03 and 0.1 anomalous data. The data can also be of a float or integer type. Each generated data is fitted into Pandas DataDrame object upon the calss' call to store the dataset as CSV. Currently I generated 1000 rows dataset until 10_000_000 rows dataset to test the POT detector. The values are also available in a shuffled version.