test(utils): ensure that t0, t1, and t2 periods are always correctly divided
T0 represents the buring days that serves as the calculation for the POT threshold (yes this is not a mistake, it is indeed the threshold of the "peak over threshold") that will define the pot data from the original dataset. T1 period serves as the value for calculating the quantile score of the anomaly score used as the threshold between anomalous and normal data. T2 period is the "time-of-interest" where the last row in T2 is always the "yesterday".