feat(detector): create peak over threshold calculation (#2)
* feat(pre-commit): update black and isort versions of pre-commit

* build(gitignore): add documentation assets and testing datasets

* build(pytest): increase passing test pecentage rule to 55%

* feat(python-version): update min. python version to 3.11.6

* ci(github-actions): update min python version to 3.11.6 and add test

* feat(models): create models/ directory for detector and notification models

* feat(detectors): create detectors/ directory for all types of anomaly detector

* feat(detector): create interface class for all anomaly detectors

AnomalyDetector is an interface class to ensure the scalability of this project e.g. after POT anomaly detector, we can create other anomaly detectors that uses otehr statistical methods.

* feat(detector): create AutoencoderAnomalyDetector class as a placeholder

AutoencoderAnomalyDetector is an anomaly detector that should implement the autoencoder method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create BlockMaximaAnomalyDetector class as place holder

BlockMaximaAnomalyDetector is an anomaly detector that should implement the block-maxima method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create DBSCANAnomalyDetector as placeholder

DBSCANAnomalyDetector is an anomaly detector that should implement the "Density-Based Spatial Clustering" method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create IsoForestAnomalyDetector as placeholder

IsoForestAnomalyDetector is an anomaly detector that should implement the "Isolation Forest" method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create MADAnomalyDetector as placeholder

MADAnomalyDetector is an anomaly detector that should implement the "Medien Absolute Deviation" method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create ZScoreAnomalyDetector as placeholder

ZScoreAnomalyDetector is an anomaly detector that should implement the "Z-Score" method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create OneClassSVMAnomalyDetector as placeholder

OneClassSVMAnomalyDetector is an anomaly detector that should implement the "1-Class Support Vector Machine" method when detecting anomalous data. For this v1.0.0, this is only a placeholder to showcase the scalability of this project.

* feat(detector): create POTAnomalyDetector for extreme value anomaly detection

POTAnomalyDetector is the main class that will be implemented in this v1.0.0. This class holds t0, t1, and t2 attributes that separates the time-periods which will be used to calculate the peak over threshold (t0), anomaly score quant score (t1), and the "time-of-interest" (t2).

* feat(types): create a custom AnomalyDetector type to enable the auto-complete

Sometimes, having 1 single class as a type, mostly if it's an interface class, will disable the auto-complete and disturb the linters because the child classes will not be visible in the background. In Python 3.12, we have the flexibility to create a simple custom types that lets us still have the accurate type for an object while being able to enjoy the auto-complete and not-so-crazy liter e.g. MyPy.

* feat(detector): create FactoryAnomalyDetector for easy initialization

This is an implementation of the "Factory" design pattern where the interface object is avoided to safe memory.

* feat(main): create the get_detector function to call the FactoryAnomalyDetector

The `detecto.py` is the main module in this package. Hence all classess and functionalities will be callable from this single file. At least that is the vision from today 18.10.2023 at 21:04:31 CET.

* feat(tests): create DataGenerator to generate simple testable data for our detectors

DataGenerator class provides the functionalities to generate data with random values from 0 to 1_000 or 10_000 or -1_000 or 10_000 to 1_000 or 10_000 enhanced with customable amount of anomaly e.g. currently I used 0.03 and 0.1 anomalous data. The data can also be of a float or integer type. Each generated data is fitted into Pandas DataDrame object upon the calss' call to store the dataset as CSV. Currently I generated 1000 rows dataset until 10_000_000 rows dataset to test the POT detector. The values are also available in a shuffled version.

* feat(tests): create conftest as the main module for testing purpose

`Conftest.py` acts as the main configuration file for the all the testing in this project. For now, it will be used to initialize the test dataframes with the previously generated data stored in `tests/datasets/**`. The test datasets themselves are not stored in GitHub but rather locally.

* test(detector): ensure FactoryAnomalyDetector returns the correct object

All anomaly detectors are already available even though some of them are only palce holder.

* test(utils): ensure that t0, t1, and t2 periods are always correctly divided

T0 represents the buring days that serves as the calculation for the POT threshold (yes this is not a mistake, it is indeed the threshold of the "peak over threshold") that will define the pot data from the original dataset. T1 period serves as the value for calculating the quantile score of the anomaly score used as the threshold between anomalous and normal data. T2 period is the "time-of-interest" where the last row in T2 is always the "yesterday".

* test(utils): ensure the correct quantile calculation for POT data

The "Peak Over Threshold" is the method that separate the normal data with the anomalous data. Only the POT data (data that are over the threshold of the POT) have the potential to be the extrme anomaly.

* test(detector): ensure compute_pot_data returns the correct pot data

`compute_pot_data()` is a method used by POTAnomlayDetector to create the POT dataset. This method utilizes the quantile calculation from `src/utils/math.py`.

* fix(tests): modify init_test_df() function to return a string if the dataset file doesnt exist

* fix(tests): add __init__.py in datasets/ dir to be noticed by git

* fix(detector): rewrite AutoEncoder into Autoencoder

---------

Co-authored-by: N. L <nino@pleno.earth>